# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary List all products
 */
export const ListProductsQueryParams = zod.object({
  "category": zod.coerce.string().optional(),
  "brand": zod.coerce.string().optional()
})

export const ListProductsResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "brand": zod.string(),
  "model": zod.string(),
  "category": zod.string(),
  "description": zod.string(),
  "specs": zod.array(zod.string()),
  "imageObjectPath": zod.string().nullish(),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
})
export const ListProductsResponse = zod.array(ListProductsResponseItem)


/**
 * @summary Create a new product (admin)
 */
export const CreateProductBody = zod.object({
  "name": zod.string(),
  "brand": zod.string(),
  "model": zod.string(),
  "category": zod.string(),
  "description": zod.string(),
  "specs": zod.array(zod.string()),
  "imageObjectPath": zod.string().nullish()
})


/**
 * @summary Get a single product
 */
export const GetProductParams = zod.object({
  "id": zod.coerce.number()
})

export const GetProductResponse = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "brand": zod.string(),
  "model": zod.string(),
  "category": zod.string(),
  "description": zod.string(),
  "specs": zod.array(zod.string()),
  "imageObjectPath": zod.string().nullish(),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
})


/**
 * @summary Update a product (admin)
 */
export const UpdateProductParams = zod.object({
  "id": zod.coerce.number()
})

export const UpdateProductBody = zod.object({
  "name": zod.string(),
  "brand": zod.string(),
  "model": zod.string(),
  "category": zod.string(),
  "description": zod.string(),
  "specs": zod.array(zod.string()),
  "imageObjectPath": zod.string().nullish()
})

export const UpdateProductResponse = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "brand": zod.string(),
  "model": zod.string(),
  "category": zod.string(),
  "description": zod.string(),
  "specs": zod.array(zod.string()),
  "imageObjectPath": zod.string().nullish(),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
})


/**
 * @summary Delete a product (admin)
 */
export const DeleteProductParams = zod.object({
  "id": zod.coerce.number()
})


/**
 * @summary Request a presigned upload URL
 */
export const RequestUploadUrlBody = zod.object({
  "name": zod.string(),
  "size": zod.number(),
  "contentType": zod.string()
})

export const RequestUploadUrlResponse = zod.object({
  "uploadURL": zod.string(),
  "objectPath": zod.string()
})